refactor(client): decompose SidePanel.tsx (#401)
Extract useAnnotationReview hook, FilterBar component, and BulkActions
component to reduce SidePanel from 911 to 504 LOC. Pure extraction —
no behavior changes, all data-testid attributes preserved.
